lib: move xref headers into include folder
