+++ include/xref/dsound.def
